+++ src/Button/Button.js
@@ -215,18 +215,6 @@ class Button extends PureComponent {
       'danger'
     ]),
     /**
-     * Если указан href, то кнопка будет ссылкой
-     */
-    href: PropTypes.string,
-    /**
-     * Атрибут target для ссылки, если указан атрибут href
-     */
-    target: PropTypes.string,
-    /**
-     * Атрибут rel для ссылки, если указан атрибут href
-     */
-    rel: PropTypes.string,
-    /**
      * Css-класс
      */
     className: PropTypes.string,
@@ -250,10 +238,6 @@ class Button extends PureComponent {
      * Размер кнопки
      */
     size: PropTypes.oneOf(['small', 'medium']),
-    /**
-     * Обработчик клика
-     */
-    onClick: PropTypes.func,
     /**
      * Элемент, который содержит контент, например `<Link />`
      * в случае с `react-router`
@@ -286,7 +270,11 @@ class Button extends PureComponent {
     /**
      * Закругленная кнопка
      */
-    rounded: PropTypes.bool
+    rounded: PropTypes.bool,
+    /**
+     * Коллбэк для передачи ссылки на ноду элемента кнопки
+     */
+    nodeRef: PropTypes.func
   }
 
   static defaultProps = {
@@ -317,7 +305,6 @@ class Button extends PureComponent {
       children,
       size,
       type,
-      href,
       container,
       buttonType,
       disabled,
@@ -329,6 +316,7 @@ class Button extends PureComponent {
       iconPosition,
       rounded,
       style = {},
+      nodeRef,
       classes, // eslint-disable-line no-unused-vars
       theme, // eslint-disable-line no-unused-vars
       ...other
@@ -372,12 +360,10 @@ class Button extends PureComponent {
 
     const resultContainer = isValidElement(container) ? (
       container
-    ) : href ? (
-      <a href={href} />
     ) : overlay ? (
-      <div />
+      <div ref={nodeRef} />
     ) : (
-      <button type={buttonType} />
+      <button type={buttonType} ref={nodeRef} />
     )
 
     return cloneElement(